refactor: 일정 facade 모듈 command/query 분리
